+++ R/methods-Spectrum.R
@@ -320,6 +320,7 @@ setMethod("isCentroided", "Spectrum",
 #' @examples
 #'
 #' ## Load a profile mode example file
+#' library(BiocParallel)
 #' register(SerialParam())
 #' library(msdata)
 #' f <- proteomics(full.names = TRUE,